Improve construction by filling missing component
Construction requires three parts a building type, contractor and a placement.
If any part was missing at the time of the request no job would be created.
Leaving many frames to rediscover the same parts till everything was available.

The process is improved by creating a Job with a building type, and later
frames can fill in the missing placement and contractor.